refactor: fetchTimestampsのビジネスロジックをTimestampServiceに分離
- TimestampServiceクラスを新規作成してタイムスタンプ取得ロジックを移動
- ChannelController::fetchTimestampsをシンプル化（約170行→約25行）
- 不要になったprivateメソッド（categorizeFirstChar、getFirstCharCategory）を削除
